Move around integration tests into appropriate files/directories
