+++ .github/workflows/ci.yaml
@@ -1,6 +1,9 @@
-name: Tests
+name: integration and other tests
+
 on:
+  workflow_call:
   pull_request:
+
 jobs:
   lint:
     name: Lint
@@ -12,6 +15,7 @@ jobs:
         run: python3 -m pip install tox
       - name: Run linters
         run: tox -e lint
+
   unit-test:
     name: Unit tests
     runs-on: ubuntu-latest
@@ -23,8 +27,8 @@ jobs:
       - name: Run tests
         run: tox -e unit
 
-  integration-test-microk8s-ha:
-    name: Integration HA tests (microk8s)
+  integration-test-microk8s-charm:
+    name: Charm integration tests (microk8s)
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
@@ -36,8 +40,8 @@ jobs:
           # This is needed until https://bugs.launchpad.net/juju/+bug/1977582 is fixed
           channel: 1.23/stable
           bootstrap-options: "--agent-version 2.9.29"
-      - name: Run HA integration tests
-        run: tox -e integration-ha
+      - name: Run charm integration tests
+        run: tox -e integration-charm
 
   integration-test-database-relation:
     name: Integration tests for database relation
