+++ .github/workflows/release.yaml
@@ -20,85 +20,14 @@ jobs:
           credentials: "${{ secrets.CHARMHUB_TOKEN }}" # FIXME: current token will expire in 2023-07-04
           github-token: "${{ secrets.GITHUB_TOKEN }}"
 
-  lint:
-    name: Lint
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Install dependencies
-        run: python3 -m pip install tox
-      - name: Run linters
-        run: tox -e lint
-
-  unit-test:
-    name: Unit tests
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Install dependencies
-        run: python -m pip install tox
-      - name: Run tests
-        run: tox -e unit
-
-  integration-test-ha:
-    name: Integration HA tests (microk8s)
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Setup operator environment
-        uses: charmed-kubernetes/actions-operator@main
-        with:
-          provider: microk8s
-          # workaround for https://github.com/charmed-kubernetes/actions-operator/issues/37
-          channel: 1.23/stable
-          bootstrap-options: "--agent-version 2.9.29"
-      - name: Run HA integration tests
-        run: tox -e integration-ha
-
-  integration-test-osm-mysql:
-    name: Integration osm-msyql relation tests (microk8s)
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Setup operator environment
-        uses: charmed-kubernetes/actions-operator@main
-        with:
-          provider: microk8s
-          # This is needed until https://bugs.launchpad.net/juju/+bug/1977582 is fixed
-          channel: 1.23/stable
-          bootstrap-options: "--agent-version 2.9.29"
-      - name: Run osm-mysql relation integration tests
-        run: tox -e integration-osm-mysql
-
-  integration-test-database-relation:
-    name: Integration tests for database relation
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Setup operator environment
-        uses: charmed-kubernetes/actions-operator@main
-        with:
-          provider: microk8s
-          # This is needed until https://bugs.launchpad.net/juju/+bug/1977582 is fixed
-          channel: 1.23/stable
-          bootstrap-options: "--agent-version 2.9.29"
-      - name: Run integration database tests
-        run: tox -e integration-database-relation
+  ci-tests:
+    uses: ./.github/workflows/ci.yaml
 
   release-to-charmhub:
     name: Release to CharmHub
     needs:
       - lib-check
-      - lint
-      - unit-test
-      - integration-test-database-relation
-      - integration-test-ha
-      - integration-test-osm-mysql
+      - ci-tests
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
