+++ tests/integration/relations/application-charm/lib/charms/data_platform_libs/v0/database_requires.py
