+++ tests/integration/relations/test_osm_mysql.py
