+++ tests/integration/test_charm.py
@@ -20,6 +20,14 @@ from tests.integration.helpers import (
     get_unit_address,
     scale_application,
 )
+from constants import CLUSTER_ADMIN_USERNAME, PASSWORD_LENGTH, ROOT_USERNAME
+from tests.integration.helpers import (
+    fetch_credentials,
+    get_primary_unit,
+    rotate_credentials,
+)
+from utils import generate_random_password
+
 
 logger = logging.getLogger(__name__)
 
@@ -31,7 +39,7 @@ CLUSTER_NAME = "test_cluster"
 @pytest.mark.order(1)
 @pytest.mark.skip_if_deployed
 @pytest.mark.abort_on_fail
-@pytest.mark.ha_tests
+@pytest.mark.charm_tests
 async def test_build_and_deploy(ops_test: OpsTest) -> None:
     """Build the mysql charm and deploy it."""
     async with ops_test.fast_forward():
@@ -73,7 +81,7 @@ async def test_build_and_deploy(ops_test: OpsTest) -> None:
 
 @pytest.mark.order(2)
 @pytest.mark.abort_on_fail
-@pytest.mark.ha_tests
+@pytest.mark.charm_tests
 async def test_consistent_data_replication_across_cluster(ops_test: OpsTest) -> None:
     """Confirm that data is replicated from the primary node to all the replicas."""
     # Insert values into a table on the primary unit
@@ -127,7 +135,7 @@ async def test_consistent_data_replication_across_cluster(ops_test: OpsTest) -> 
 
 @pytest.mark.order(3)
 @pytest.mark.abort_on_fail
-@pytest.mark.ha_tests
+@pytest.mark.charm_tests
 async def test_scale_up_and_down(ops_test: OpsTest) -> None:
     """Confirm that a new primary is elected when the current primary is torn down."""
     async with ops_test.fast_forward():
@@ -172,5 +180,95 @@ async def test_scale_up_and_down(ops_test: OpsTest) -> None:
         ]
         assert len(not_online_member_addresses) == 0
 
-        await scale_application(ops_test, APP_NAME, 0)
-        await ops_test.model.remove_application(APP_NAME)
+
+@pytest.mark.order(4)
+@pytest.mark.abort_on_fail
+@pytest.mark.charm_tests
+async def test_password_rotation(ops_test: OpsTest):
+    """Rotate password and confirm changes."""
+    random_unit = ops_test.model.applications[APP_NAME].units[-1]
+
+    old_credentials = await fetch_credentials(random_unit, CLUSTER_ADMIN_USERNAME)
+
+    # get primary unit first, need that to invoke set-password action
+    primary_unit = await get_primary_unit(ops_test, random_unit, APP_NAME)
+    primary_unit_address = await primary_unit.get_public_address()
+    logger.debug(
+        "Test succeeded Primary unit detected before password rotation is %s", primary_unit_address
+    )
+
+    new_password = generate_random_password(PASSWORD_LENGTH)
+
+    await rotate_credentials(
+        unit=primary_unit, username=CLUSTER_ADMIN_USERNAME, password=new_password
+    )
+
+    updated_credentials = await fetch_credentials(random_unit, CLUSTER_ADMIN_USERNAME)
+    assert updated_credentials["password"] != old_credentials["password"]
+    assert updated_credentials["password"] == new_password
+
+    # verify that the new password actually works
+    # since get_primary_unit (and this get_cluster_status) use the cluster admin credentials
+    primary_unit = await get_primary_unit(ops_test, random_unit, APP_NAME)
+    primary_unit_address = await primary_unit.get_public_address()
+    logger.debug(
+        "Test succeeded Primary unit detected after password rotation is %s", primary_unit_address
+    )
+
+
+@pytest.mark.order(5)
+@pytest.mark.abort_on_fail
+@pytest.mark.charm_tests
+async def test_password_rotation_silent(ops_test: OpsTest):
+    """Rotate password and confirm changes."""
+    random_unit = ops_test.model.applications[APP_NAME].units[-1]
+
+    old_credentials = await fetch_credentials(random_unit, CLUSTER_ADMIN_USERNAME)
+
+    # get primary unit first, need that to invoke set-password action
+    primary_unit = await get_primary_unit(ops_test, random_unit, APP_NAME)
+    primary_unit_address = await primary_unit.get_public_address()
+    logger.debug(
+        "Test succeeded Primary unit detected before password rotation is %s", primary_unit_address
+    )
+
+    await rotate_credentials(unit=primary_unit, username=CLUSTER_ADMIN_USERNAME)
+
+    updated_credentials = await fetch_credentials(random_unit, CLUSTER_ADMIN_USERNAME)
+    assert updated_credentials["password"] != old_credentials["password"]
+
+    # verify that the new password actually works
+    # since get_primary_unit (and this get_cluster_status) use the cluster admin credentials
+    primary_unit = await get_primary_unit(ops_test, random_unit, APP_NAME)
+    primary_unit_address = await primary_unit.get_public_address()
+    logger.debug(
+        "Test succeeded Primary unit detected after password rotation is %s", primary_unit_address
+    )
+
+
+@pytest.mark.order(6)
+@pytest.mark.abort_on_fail
+@pytest.mark.charm_tests
+async def test_password_rotation_root_user_implicit(ops_test: OpsTest):
+    """Rotate password and confirm changes."""
+    random_unit = ops_test.model.applications[APP_NAME].units[-1]
+
+    root_credentials = await fetch_credentials(random_unit, ROOT_USERNAME)
+
+    old_credentials = await fetch_credentials(random_unit)
+    assert old_credentials["password"] == root_credentials["password"]
+
+    # get primary unit first, need that to invoke set-password action
+    primary_unit = await get_primary_unit(ops_test, random_unit, APP_NAME)
+    primary_unit_address = await primary_unit.get_public_address()
+    logger.debug(
+        "Test succeeded Primary unit detected before password rotation is %s", primary_unit_address
+    )
+
+    await rotate_credentials(unit=primary_unit)
+
+    updated_credentials = await fetch_credentials(random_unit)
+    assert updated_credentials["password"] != old_credentials["password"]
+
+    updated_root_credentials = await fetch_credentials(random_unit, ROOT_USERNAME)
+    assert updated_credentials["password"] == updated_root_credentials["password"]
