+++ tox.ini
@@ -65,8 +65,8 @@ commands =
         -m pytest --ignore={[vars]tst_path}integration -v --tb native -s {posargs}
     coverage report
 
-[testenv:integration-ha]
-description = Run integration tests
+[testenv:integration-charm]
+description = Run charm integration tests
 deps =
     juju==2.9.11 # juju 3.3.0 has issues with retrieving action results
     mysql-connector-python
@@ -75,7 +75,7 @@ deps =
     pytest-order
     -r{toxinidir}/requirements.txt
 commands =
-    pytest -v --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs} -m "ha_tests"
+    pytest -v --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs} -m "charm_tests"
 
 [testenv:integration-database-relation]
 description = Run database relation tests
